Fix docs root redirect
